BUG: qr_delete failures in economic mode.
The primary bug was using statically allocated arrays of size 1
instead of size 2.

A second bug was in the failure path for
the call to reorthx.  When deleting a row, we really just need
a vector into the null space of Q.  It can be derived from the
kth basis vector, (as in the first call to reorthx) in which case,
we need the length of the vector into the null space, (s[n] in
reorthx and reorth and rho in Reichel's paper), however if we
don't use the kth basis vector when deleting the kth row, s[n]
should be 0, but s[:n] should be the values associated with
the null space vector we are using.  This is discussed in
Daniel p.785.  Reichel doesn't discuss this at all.

Also fixed the inv_root2 to actually be
1/sqrt(2) instead of sqrt(2) in reorthx and reorth.

fixes the qr updating part of gh-4885

(cherry picked from commit e08a77a10dbdac06d77bdcfddd3afcdec9eea930)