MAINT: Move distance_p to MinMaxDist
side_distance_from_min_max used by query.cxx
is actually one of the interval functions in MinMaxDist.
We need to find a way of porting query.cxx to that.
But since there has been quite some efforts in optimizing
the speed of query.cxx I hesitate on this.

Go with the other queries first.