MAINT: assimilate p=inf case.
This also fixed a few bugs in the periodic case of last commit.

Also revealed is the non-periodic test cases are not as powerful
as the periodic test cases. They couldn't find it if the min/max reduction
in MinMaxDist is wrongly typed as *max = dmax(*min, min_);

Maybe we shall switch to a uniform distribution in the testcases.