ENH: Treat P-norms with template.
This made the code slightly faster and leaner by moving most of
the p-norm branching in distance calculation from the deep down
to the top of the query_xxx function calls (via these HANDLE macros)

All four cases INF, 2, 1, p are handled.

Note that query.cxx is not yet ported to this new infrastructure.