Fix misleading wavelets/cwt example
extent is in the wrong order. if used as before the change, the
resulting image suggests that the best matching ricker wavelet
width (i.e. it's scale) is in (27, 28).
The widths, however, are related to the number of samples on the
time axis - in the examples case the time axis is 200 entries long
spanning from (-1, 1) seconds and contain a 7Hz signal.
This amounts 1 second being represented by 100 points on the t axis,
which - in turn - means a period covers 14.28 dots on the t axis.
According to the defintion of ricker wavelet the optimal scale
to match this signal would be 0.25 * period lenght or - in
this case - 3.57 as opposed to between 27 and 28

Reversing the axis conveys exactly this putting the best match
between 3 and 4