BUG: dendrogram: remove color_list kw from signature, it doesn't belong there.
It wasn't in the parameters list and obviously wasn't meant to be there, given
that (a) it is initialized as an empty list at the top of the function, and
(b) the length of the input needed is hard to guess.

Note that there is a function to set the color palette for dendrogram, namely
`set_link_color_palette`.

Closes gh-4926.