BUG: special: Rewrite the Fortran code that computes the zeros of Y'.
In the issue gh-4690, it was reported that jn_zeros(281, 6) never returned,
and the python process would have to be killed.  Also, ynp_zeros(280, 4)
and ynp_zeros(282, 4) returned zeros that were not consecutive.

The subroutine JYZO in specfun.f computes the zeros of J, J', Y and Y' using
Newton's method.  This is the function used by jn_zeros, jnp_zeros, yn_zeros
and ynp_zeros.  The problem was that the code that computed the zeros of Y'
did not always provide a good initial guess for Newton's method.  When the
code saw that Newton's method had converged to a previously found zero of Y',
it added pi to the initial guess and tried again.  This did not always work,
and it didn't prevent the code from converging to a zero that was larger than
the expected zero.  Neither Newton's method nor the retry loop were protected
from iterating indefinitely. In the particular case of jn_zeros(281, 6),
eventually it used an initial guess for a zero of Y' that put Newton's method
into an infinite loop.

The fix made here is to use the previously computed zeros of J as the initial
guesses of the zeros of Y', and then only allow at most 10 Newton iterations.
(After that, NAN is returned.)  This works very well.  I haven't found a case
where Newton's method fails to converge.

Closes gh-4690