BUG: Don't pass garbage strides to BLAS.
When inserting columns or updating QR decompositions of (1,1) or
(1,N) matrices nonsensical stride data could be passed to ?gemv.
This caused the sgemv wrapper for OSX to error.  (1,1) and (1,N)
systems are now special cased.

This commit also removes testing form_qTu for q (1,1).  Form_qTu
is not ever called with q (1,1) from this commit.

Closes gh-5149.