MAINT: add a replacement inspect.getargspec for python 3.x
This is borrowed from Django,
https://github.com/django/django/pull/4846

Note an inconsistency between inspect.getargspec(func) and
inspect.signature(func). If `func` is a bound method, the latter does *not*
list `self` as a first argument, while the former *does*.
Hence cook up a common ground replacement: `getargspec_no_self` which
mimics `inspect.getargspec` but does not list `self`.

This way, the caller code does not need to know whether it uses a legacy
.getargspec or bright and shiny .signature.