BUG: io: Stop guessing the data delimiter in ARFF files.
In the ARFF reader, there were several dozen lines of code that
determined whether the delimiter in the @DATA section of the ARFF
file was a comma or a space.  This code has been removed.  The ARFF
file format specification says the delimiter must be a comma.

As a side effect, this closes gh-5276.  'loadarff' can now handle
a file with no data in the @DATA section.