BUG: misc: Fix the Pillow work-around for 1-bit images.
Pillow has a bug that can cause a seg. fault when a 1-bit
image is read and converted to a numpy array.  Previously,
code was added to convert an Image instance to 8-bit before
converting it to a numpy array, but this change didn't correctly
convert the image to 8-bit (the convert method does not operate
in-place).

Closes gh-2138.