DOC: special: Remove a potentially misleading example from the diric docstring
The final part of the last example in the diric docstring shows a calculation
that produces the same result as an FFT.  The example is misleading, because
it only works in special cases. (E.g. it doesn't work if k is changed from 3
to 2.)  This commit removes that part of the example.

[ci skip]