EHN: Precompute v5 savemat's byte counts
`VarWriter5.writetop` should precompute the byte count of the array
using its current framework before saving it, allowing the write to be
streamed instead of having the byte count written out of order.

This removes the need for calls to update_matrix_tag, as the matrix tag
only needs to be written once per item in the array instead of twice.