BUG: Fix OverflowError for saving large matrices
The main problem causing "OverflowError: size does not fit in an int"
is an implementation issue of the zlib module of python 2.7. A simple
way to resolve this is to simply use the compression object that
the zlib module also includes to compress data parts at a time. Even
in the case that a single write exceeds 2GB, the string can be split
into multiple parts for the write.