ENH: Add zero_phase kwarg to signal.decimate
Add the capability for ``decimate`` to use ``filtfilt`` instead of ``lfilter``
when using an IIR downsampling filter, to eliminate the generally unwanted
phase shift between the input and output data. Tests have been adapted and
extended to account for this behavior. A warning is now given when the
`zero_phase` kwarg is not set, as the default will be `True` in a future
release.

Adapted from https://gist.github.com/FRidh/ba60084570d05c91e9bd

Closes #4929