_Qhull should raise ValueError for NaNs.
ConvexHull would trigger a Segmentation Fault, which is not a very
elegant way to show that something is wrong. As it should not try to
figure something out of NaNs, it beautifully resigns with a ValueError
now.

This also reflects on the behavior of Delaunay.

Tests that would pass if they didn't have NaNs have been written for
both classes.

Fixes #5450