ENH: make optimize.{bisect,brentq,brenth,ridder} have consistent tolerance
Given an initial interval [a, b], an absolute tolerance xtol, and a
relative tolerance rtol, the methods currently achieve the following
tolerances:

- bisect, ridder: |x - x0| < xtol + rtol (|a| + |b|)
- brentq, brenth: |x - x0| < 2(xtol + rtol |x0|).

Here x0 is the computed root and x is the exact root. This commit
changes the functions so that they all achieve a tolerance of

|x - x0| < xtol + rtol |x0|;

in other words they should satisfy ``np.allclose(x, x0, atol=xtol,
rtol=rtol)``. It also changes the tests to check for this criterion
and documents the tolerance used. Closes gh-5555.