Fixed up sigmaclip function - No axis argument
Can't figure out a decent way to keep the axis argument - So dropping it for
the moment.

Added tests for using a custom `cenfunc` in sigmaclip as well as using the
`iters` keyword.

Added my name to the THANKS.txt file.

Note: It seems that the "new" sigmaclip funcion is faster than the old
"sigmaclip" function - This is from some benchmarks I've run in IPython.
I don't believe that the new code can be faster, because the basic algorithm
hasn't been touched.

There is an obvious speedup to be gained if you use the `iters` keyword but the
speedup I've noticed is wih `iters = None`. So not sure what to make of that.