BUG: eliminate warnings and (on Windows) exception in test_maskandscale().
The primary change is that the variable referring to a memory map
is now deleted before attempting to close the file.

In addition, the presence of nan behind the mask in a masked array
was triggering a warning when the array was divided by the scale
factor.  This might be viewed as a bug in ma; the workaround here
is to use a large value instead of nan.