BUG: io.netcdf masking requires exact match
The masking being done by io.netcdf, upon reading a variable, was
masking values that were approximately equal to the _FillValue. This
commit fixes the behavior so that only exact matches are masked.

This also adds a test that demonstrates the behavior; this test fails
without the modification to netcdf.py.

Fixes ticket #5626