TST: fix io.netcdf test that was failing CI tests
The python3.5 test was failing with the message below. I believe that
this is a problem in the test, rather than in netcdf.py itself. I am
working around this by opening the file with mmap=False for this
test. It should also work to delete vardata after the assertions (which
seems to be done elsewhere in test_netcdf.py), but I'm concerned that
that would lead to problems if the assertions failed (in which case I
think the 'with' block would exit before deleting vardata).

I am unable to reproduce this problem on my machine, so I'm submitting
this commit in order to see if the CI tests now pass.

ERROR: test_netcdf.test_read_withMaskAndScaleFalse
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/travis/virtualenv/python3.5.0/lib/python3.5/site-packages/nose/case.py", line 198, in runTest
    self.test(*self.arg)
  File "/home/travis/virtualenv/python3.5.0/lib/python3.5/site-packages/scipy/io/tests/test_netcdf.py", line 417, in test_read_withMaskAndScaleFalse
    assert_equal(vardata, [1, 2, 3])
  File "/home/travis/virtualenv/python3.5.0/lib/python3.5/site-packages/scipy/io/netcdf.py", line 306, in __exit__
    self.close()
  File "/home/travis/virtualenv/python3.5.0/lib/python3.5/site-packages/scipy/io/netcdf.py", line 297, in close
    ), category=RuntimeWarning)
RuntimeWarning: Cannot close a netcdf_file opened with mmap=True, when netcdf_variables or arrays referring to its data still exist. All data arrays obtained from such files refer directly to data on disk, and must be copied before the file can be cleanly closed. (See netcdf_file docstring for more information on mmap.)