ENH: Reduce operation count in rev. com iterative solvers
The reverse communication interface used by most of the iterative
solvers is built around the dgemv interface:

y <- alpha * A(x) + beta*y

And provides both alpha and beta scalars.  The linear operator
interface we have doesn't directly support this operation.  If
alpha and beta have special values (1 and 1, or -1 and 0, etc)
this operation can be completed without performing as many
floating point operations.  This commit removes a number of these
redundant floating point operations.

Here this provides a 10-20% reduction in execution time.