BUG: fix generator in io._loadarff to comply with PEP 0479
In the function _loadarff a generator is created which yields the data
from an arff file. If given a file with no data, it tries to yield
after EOF, which raises StopIteration and violates PEP 0479. This
commit moves all of the parsing inside a for loop which correctly
handles the no-data file case. Closes gh-5673.