ENH: circvar calculated simply as 1-R
The circular statistic function circvar is updated to estimate circular
variance simply as one minus the mean resultant vector (1-R), instead
of using approximation to the linear variance. This definition is
compatible with the literature (e.g., Fisher, 1993; Jammalamadaka and
SenGupta, 2001) and other implementations in statistical packages (e.g.,
R and Matlab).

Optional boolean parameter ``normalize`` added to both circvar and
cirstd determines whether the returned values are independent on the
variable units, or whether the values are scaled by the range of
variable values (currently default not to break compatability).

Docstring of circvar, circstd and circmean have more accurate
description for the parameters ``high`` and ''low``.