BUG: Enforce 'x' input to Data init as array_like
Previously, it was possible to pass in scalar values
to 'x' without error for 'Data' objects. However, it
renders such data unusable in 'ODR' because it throws
a seemingly inexplicable 'IndexError' during initialization.

This commit causes scalar inputs for 'x' to throw a 'ValueError'
during 'Data' object initialization, and there is now a test in
'test_odr.py' to check that that remains enforced.