MAINT: warn user if provided data is empty to ODR
When empty data is passed into 'ODR', fitting will
just return the original provided coefficients / data,
which is not particularly meaningful. While this problem
is already addressed in the 'odr' module, surfacing it
as a warning will make it clearer to the user that passing
in empty data isn't a great idea.

Closes gh-5721.