BUG: Ensure scipy's _asfarray returns ndarray
Enforces ndarray return type for scipy.fttpack.
basic._asfarray, for having a 'dtype' attribute
does not guarantee that the passed in object was
an ndarray. A prominent exception is the Series
object from the pandas library.

When "array", non-ndarray objects like these are
passed into methods like fftpack.rfft, the data
is overwritten as a result, which is not what
we want.

Closes gh-5790.

Addendum: The "ci skip" is added below because the
previous version of the commit passed all unit
tests, and this current version incorporated a comment
change via rebase, so there was no need to rerun tests.

[ci skip]