BUG: sparse.linalg: add locks to ensure ARPACK threadsafety
The ARPACK fortran code uses SAVE variables, and is not threadsafe or
re-entrant. Add locks and reentrancy checks to the Python side to deal
with this issue --- previously the code used to segfault.

As we use an explicit lock, it is not necessary for the Fortran code to
hold the GIL, so adjust the .pyf file accordingly.

Fixes gh-5846