Returns section at interp2d class description
Current description of interp2d "Returns" seems very misleading, since
interp2d is actually a class and not a method or a function. Thus,
calling class constructor will obviously return an instance of the
class, which in this case is a callable object.

In addition, this "Returns" description content seems out of place,
since it doesn't have anything to do with a 2-D interpolation.

Maybe I'm just not getting it right. In that case, please excuse me for
bothering you.

Cheers.