Fix precision issues in sparse vs. dense float matrix sum comparisons.
Comparisons between the results of the sparse and dense matrix `.sum`
methods in `sparse.test_base.test_sum` fail for matrices containing
single-precision floats (`np.float32` and `np.complex64`) when the
sparse and dense sums are calculated differently and the resulting sum
values have too many digits to be considered equal in single-precision
float format with the default value of the `decimal` parameter of
`assert_array_almost_equal`. This is fixed by using smaller test
matrices, so that the number of digits of their sums remains small
enough in all cases for the values to be considered equal at the
default value of the `decimal` parameter.