ENH: speed up 'full' and 'valid' 1D convolution
np.convolve doesn't care about input size in 'full' or 'valid' mode, so
use it.

In 'same' mode, np.convolve uses the size of the larger input, while
scipy uses the size of the first input, so if the first is also the
larger, it's ok to use np.convolve.