Update info/URL for octave-maintainers discussion
MATLAB array class 17 (0x11) still lacks documentation. The reference to a useful octave-matainers discussion in a comment of this source file was broken.  I've tracked it down their new list hosted at lists.gnu.org.  The reason for the two URLs is because the thread is spread out in their archive; the first URL provide the original post and the second the informative discussion.