MAINT: Replace calls to `hanning` with `hann`
This is just a nitpick over internal calls to the hann window. Users who
request `hanning` still get the expected result.

Quoting the hann window docsctring:
```
The window was named for Julius von Hann, an Austrian meteorologist. [...] It
is sometimes erroneously referred to as the "Hanning" window, from the use of
"hann" as a verb in the original paper and confusion with the very similar
Hamming window.
```