MAINT: further improve accuracy around 0, 1, and 2.
Expanded the range of the Taylor series around 1 to a disk of radius
1/2 by using an explicit expression in terms of the zeta
function. Started computing the function around 0 and 2 by using the
recurrence relation to shift to 1 and then using the Taylor series.
This increases accuracy and simplifies the code. There are now no
known regions in the complex plane where the relative error is greater
than 1e-13.