BUG: try to fix Travis failures.
The added tests show that the accuracy around z = 1 is fine, which
makes it likely that the Taylor series is fine. The accuracy has
problems around z = 2. Results in that region are computed by the
formula log(z - 1) + loggamma(z - 1). The loggamma term is computed by
the Taylor series, so differences in accuracy for log around z = 1 on
different systems are the prime suspects. Add an expansion of log
around 1 to test this hypothesis.