MAINT: reimplement special.digamma for complex arguments
The old digamma function for complex arguments was a known failure;
for comparison:
- on the new TestSystematic test it had 718/3779 failures
- on the new test around the negative real axis (test_digamma_negreal)
it had 990/1100 failures.
The new version has 0. Note that on one boost test the new function
does worse than the old one; at a single point it acheives a relative
error of about 2e-15 whereas the old one was under 1e-15. The test
point was real, however, and the old complex digamma function handed
that case to Cephes. I bumped the tolerance to 1e-14 to get that test
to pass.

This commit also improves the accuracy of both the real and complex
digamma functions around the zeros closest to the origin.

Finally, this commit adds internal functions _sinpi and _cospi. These
show up frequently in reflection formulas (currently for loggamma and
digamma) so this enables reuse.