MAINT: improve special.gammainc in the ``a ~ x`` regime.
Before gammainc would hang in the ``a ~ x`` regime because the methods
used to compute it converged very slowly. This adds an asymptotic
series tailored for that region.

It also adds a folder _precompute to special which contains the code
used to precompute the coefficients of the asymptotic series as well
as tests to verify that these coefficients are correct. The idea is to
ensure reproducibility and aid debugging.