BUG: stats: Fix broadcasting in the rvs() method of the distributions.
The class rv_generic was updated to handle broadcasting of the
arguments to the rvs() method.

The following distributions required additional custom changes to
support broadcasting in rvs():
    levy_stable, pearson3, rice, truncnorm,
    hypergeom, planck, randint.

The function `numpy.broadcast_to` is used in several places.  This
function was added to numpy in version 1.10, so a copy was added to
_lib/_numpy_compat.py to support older version of numpy.

Closes gh-2069.