BUG: optimize/slsqp: don't overwrite an array out of bounds
Although the array `y` is declared in the subroutine to be of size
`m+2*n`, the actually allocated size on line 251 before call to SLSQPB
is smaller.  Before db343b85053f, only `m+2*n3` multipliers were copied
into y, so change the nan writing to match that.  Add a test that
triggers the condition n3 < n.