ENH: stats: Use explicit formulas for sf, logsf, etc in weibull distributions.
The functions _sf and _logsf are added to the frechet_r/weibull_min
distribution.

The functions _logpdf, _logcdf and _sf are added to the frechet_l/weibull_max
distribution.

The main advantage of these additions is these distributions will now
give more accurate values for the survival function where the default
formula of 1 - CDF gives 0.