DOC: Add some references and optimization details
Harris on Blackman: "As an observation, note
that the coefficients of the Blackman window sum to zero
(0.42 -0.50  +0.08) at the boundaries while the exact coefficients
do not.  Thus the Blackman window is continuous
with a continuous first derivative at the boundary and falls off
like 1/w^3  or 18 dB per octave.  The exact terms (Like the
Hamming window) have a  discontinuity at the boundary and
falls off like 1/w or 6 dB per octave."