MAINT: replace author name map in tools/authors.py with git .mailmap
Git has a standard format for author name/email aliases, so use that
instead of a custom approach.

The information used to generate the .mailmap came from 'git log' and
the current contents of tools/authors.py (which was user-provided
information). Both of these are information publicly provided by the
authors for the purpose of identifying them.