MAINT: stats: remove the cython implementation of ks_2samp
Since it's no longer used. It was added in gh-5938 for scipy 0.18.0
to get some speedup for ks_2samp, but then the addition was reverted
in gh-6545, following the discussion in gh-6435: it gives different
answers on different machines, it changes one ad hoc statistic to
a different ad hoc statistic, and neither of them are clearly "correct".