MAINT: linalg: make handling of infinite eigenvalues in `ordqz` explicit
When there are infinite eigenvalues the sorting functions in `ordqz`
end up dividing by zero. This PR modifies the sort functions to handle
the infinite eigenvalue case explicitly, adopting the convention that
infinite eigenvalues do not lie in the left-hand plane or the
right-hand plane, but they do lie outside the unit circle.

This silences some zero division warnings and resolves an ambiguity
for left-hand and right-hand plane sorting: before the eigenvalue
-1.0/0 was considered to lie in the left-hand plane and the
eigenvalue 1.0/0 was considered to lie in the right-hand plane despite
the fact that both correspond to the complex point at infinity.