TST: Spectral functions' window correction factor
It was brought to light in gh-6565 that the validity of the power and
amplitude window correction factors in `signal._spectral_helper` were
not being tested for `'spectrum'` and `'density'` modes, respectively.

This adds a test that checks that the power spectrum peak height and
integrated power spectral density of a sinusoid are (mostly) invariant
when using different window functions.