BUG: CS matrix prune should copy data from larger arrays
Pruning resizes the "data" and "indices" arrays after explict zeros
have been removed from a sparse matrix. However, this is implemented
using slicing, which just returns a view of the original array. If
the original array is relatively large, a new array should be used
instead of a view, allowing the original to be garbage collected.