BUG: optimize: add sufficient descent condition check to CG line search
Polak-Ribiere CG with inexact strong Wolfe line search does not
guarantee pk points to descent direction (unlike Fletcher-Reeves).

This makes the algorithm bail out with warnflag/non-success set if the
step direction is a non-descent direction, so that some problems that
should in principle be solvable become non-solvable.

Fix the CG line search by adding a sufficient descent condition check,
based on discussion in Gilbert & Nocedal, SIAM J. Optimization 2, 21 (1992).

The MINPACK line search does not seem to be as easily modified, so add
the check to the pure-Python search. The condition is probably not very
common, so the performance is not as important.