The correct axis to sum w*f over is the last one
Since w has shape (n,), the result of multiplication w*func(y, *args) will have the _last_ dimension equal to n, and this is the dimension we want to sum over. There may be other dimensions if f is vector-valued; summing over axis=0 is incorrect in this case. The change enables using fixed_quad for integrating vector-valued functions, and has no effect on integration of scalar-valued functions.