ENH: calculate population energies in DifferentialEvolutionSolver.__enter__
This makes it possible to drop a couple of checks for `np.inf` values
during solver startup or iteration and it makes sense to reset the
population in the __enter__ method anyways.

In tests, rename test_population_initiation into test_invalid_init_param
as other checks (not very useful by the way) are dropped as they do not
apply anymore.