cdf2pdf helper function
Often I must convert continuous density functions to probability density functions. While it is not too hard to do by hand, it would be convenient to have this function in `scipy`. `stats.py` seemed like the right place to put this, but I am open to suggestions. If there is interest, I can add documentation changes and error checking to this PR.