BUG: Fixed bug caused by discrepancy in treatment of NaNs accross platforms
On some platforms, using NaN as an operand of < generates an error, on others only a warning. The warning allowed NaNs to be detected later on most systems and allow for a graceful failure; the error crashed the routine when the problem included redundancies, presolve was turned off, and sparse solver was in use. This bug is (hopefully) fixed although I cannot test on my own machine.