MAINT: special: Rewrite a test of wrightomega
assert_equal(res, 0) fails with the development version of numpy,
because assert_equal() now takes into account the signs of the
real and imaginary parts of +/- 0.0 +/- 0.0j.

To avoid that, the test now uses explicitly given expected results
with signed real and imaginary parts, and compares the actual results
with the expected results using assert_equal() with only real arguments.
This works for old numpy and should continue working with new versions
of numpy.

Closes gh-7227.