FIX 7262: ckdtree crashes in query_knn.
This was because a loosen check on j >= 0 in the branches. The loop
was not carefully written. It has been replaced with a version that
is easier to understand.

Also added the docstring warning that if k is not sorted the result
is undefined. We can worry about adding a sorting wrapper in cython
later if there is a demand.

This fixes #7262

Use sturla's O(n) algorithm.

Also move the missing neighbour logic from Cython to C++.

Backport of gh-7266