Bug fix in resample().
The current version of resample(x,n) fails when the smaller of x.size and n is even. With real-valued x, the problem appears only on downsampling (for mathematical reasons), with complex-valued x, both on upsampling and downsampling.
Simple test cases:
1. For x=[1.+0.j, 0.+0.j], resample(x,4) should produce [1.+0.j, 0.5+0.j, 0.+0.j, 0.5+0.j], but does not.
2. For x=[1., 0.5, 0., 0.5], resample(x,2) should produce [1., 0], but does not.
The problem lies with the Fourier component at frequency N/2. Say we want to upsample from an even N to a larger num. The current code takes the single Fourier component at N/2 in the input spectrum and assigns it entirely to num-N/2 in the output spectrum. Instead that should be split evenly between frequencies N/2 and num-N/2. Vice versa say we want to downsample from Nx to a smaller N that is even. The current code takes the Fourier component at Nx-N/2 in the input spectrum and assigns it to N/2 in the output spectrum. Instead we need to add to that the Fourier component in the input spectrum at N/2.
The revision treats these special cases properly.