Addresses #7456, #7426, #7455 (scipy.special.smirnov, scipy.stats.ksone)  scipy.special.{smirnov,smirnovi} have accuracy & convergence issues  incorrect values for scipy.stats.ksone.pdf()
Add a smirnovp(), derivative of smirnov(), function.
Use smirnovp() in smirnovi() instead of an asymptotic.
If the initial estimate for x in smirnovi() is not going to work
 for Newton-Raphson (because it is on a "bad" part of the curve),
 perform some bisection steps to get a better estimate, one which
 is on the correct side of the desired value.
Replace the relative stopping criteria with same logic used in np.allclose().
With these changes, the algorithm seems to converge in
 about 5 NR iterations (10 function evaluations) for most values of p.
Exposing smirnovp() allows its use in scipy.stats.ksone() which addresses
 the inaccurate pdf(). Requires changes in many cephes infrastructure files,
 adding doc, exports.