MAINT: Avoid using module wide setup during tests
For most tests they never seemed very important, the linalg/arpack
test did not seem to actually give the warning (for me) so removed
it completely. The messiest one is the sparse tests, which requires
a lot of filtering.

Avoiding the module wide setup removes a python 3 deprecation warning
caused by nose (this is filtered by newer numpy versions, but shows
up when testing with an older version).