MAINT: Pytest testing framework setup
Pytest configuration differs in some important aspects from
numpy.testing.  It is in practice not really possible to configure it
programmatically inside the tested package --- for example conftest.py
run too late to be able to add command line arguments.

It appears difficult to fight this, and the testing approach has to be
adapted.

In particular:

- We provide a thin scipy.test() wrapper that calls pytest.main with
  a reasonable choice of command-line options.

- Warning filters etc. are configured in the pytest.ini.  These settings
  will only affect the development environment, and will be absent when
  the user tests the scipy package.

- The directory where you run pytest is important, because it affects
  where it looks for the configuration files!

The "develop" and "release" mode warning filters are then specified
implicitly.