scipy.io.savemat: handle objects supporting PEP 519
Similar logic to the previous changes to scipy.io.loadmat: first check whether
the `file_name` object has a 'write' method. If so, use it as-is. Otherwise,
append '.mat' if possible and appropriate, and pass `file_name` to `open`.
This allows the user to give any object to `savemat` which `open` understands,
including `pathlib.Path` and `os.DirEntry` under Python ≥ 3.6.