DOC: add a SciPy Code of Conduct.
This code of conduct has been discussed on the mailing list so far.
It starts with an adaptation of the Apache CoC.
The later bits about reporting and handling reports are derived from
the Contributor Covenant and the Jupyter CoC.

DOC: update Code of Conduct for review comments on gh-7963.

Version of email text on urgent interventions

I added a rewritten and more specific version of the text I sent to the
mailing list today, as discussed on the hangout.

I also tried to address Nathaniel's point about the importance of
addressing subtle discrimination, particularly against women.  I wasn't
sure what to add to the procedure, but I thought at least it was worth
emphasizing that we are aware of the problem, and that we intend the
procedure to apply.

DOC: update CoC for last review comments on gh-7963.

[ci skip]

(backport of gh-7963)