MAINT: Avoid using older np.ndarray[ndim=?] syntax.
Usually a memoryview object is preferred as the support in current
Cython is better on memoryview than the old numpy array syntax.

In addition, in most of the cases we are dealing with C-contiguous
array allocated internally , so we can deterministicly compute
the address of the buffer anyways directly using .data element.

This fixes 7829 by avoiding a potential bug in cython regarding
these older nupmy array syntax accesses.