DOC: update LICENSE.txt with licenses of bundled libs as needed.
In summary, all bundled libraries that are BSD or similar licensed
are listed, and links to their included license files are given.
Public domain software is not listed, because that is not required.

Followed the format of Debian:
https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/

Triggered by gh-7093.  Licenses for libraries put into wheels
should not be put in this file, but added to it when building the
wheels.  That part is still to be addressed, probably in multibuild.

[ci skip]