BUG: special/cdflib: fix fatal loss of precision issues in cumfnc
Replace gammaln(a+b)-gammaln(a)-gammaln(b) -> -betaln(a,b)
to avoid loss of precision in cancellations for large a, and
make use of gammaln(a+1) == gammaln(a) + log(a) for a > 0.

Whether the previous code worked correctly or not for very large values
depends on floating point issues --- it is dangerous to rely on
gammaln(a) and gammaln(a+b) returning exactly the same floating point
value when a==a+b up to fp precision, as this is apparently not
guaranteed by gfortran optimizations.