ENH: re-enabled old sv lapack routine as defaults
This fixes the speed regression in #7847. Since the default
change of sv to svx the solve routine suffered a huge performance
penalty for anything but low order NRHS.

This commit fixes that issue by enabling the use of either 1) the
svx or 2) the sv routines. With a default on the latter.

However, the main funtionality by using svx was the easy check
of the condition number to assert a non-singular matrix.
This commit adds a call to the con routines to extract the
appropriate condition number. This forces the addition of:

   - lamch (machine precision extraction)
   - gecon (condition number calculation from LU factorization)
   - lange (1/I-norm of matrix)

This commit adds to arguments (and deprecates one) by:
  - refine, bool, decides whether the svx (true) or sv (false)
    routines should be used.
  - form = 'none', 'trans', 'conj' which refers to the form
    of the solution step. This is only valid if assume_a='gen'.
    A ValueError will be issued if not 'gen'.
    This keyword deprecates the "transposed" boolean value which
    only allowed 'T' (real) and 'C' (complex) but not 'T' for
    complex.

A fix for the deprecated sym_pos keyword which wasn't used in the previous
version. This has sadly resulted in two delocalized tests failing because
sym_pos, see the OptimizeWarning for scipy/optimize/tests/test_linprog.py.

A couple of additional tests have been added, mainly to check the arguments.

Signed-off-by: Nick Papior <nickpapior@gmail.com>